Refactor DuplicateTemplateView
